+++ src/Bartlett/CompatInfoDb/ExtensionFactory.php
@@ -29,10 +29,11 @@ class ExtensionFactory implements ReferenceInterface
     const LATEST_PHP_5_3 = '5.3.29';
     const LATEST_PHP_5_4 = '5.4.45';
     const LATEST_PHP_5_5 = '5.5.38';
-    const LATEST_PHP_5_6 = '5.6.39';
+    const LATEST_PHP_5_6 = '5.6.40';
     const LATEST_PHP_7_0 = '7.0.33';
-    const LATEST_PHP_7_1 = '7.1.25';
-    const LATEST_PHP_7_2 = '7.2.13';
+    const LATEST_PHP_7_1 = '7.1.26';
+    const LATEST_PHP_7_2 = '7.2.14';
+    const LATEST_PHP_7_3 = '7.3.2';
 
     protected $storage;
 
@@ -167,7 +168,10 @@ class ExtensionFactory implements ReferenceInterface
         if (version_compare($phpVersion, '7.2', 'lt')) {
             return self::LATEST_PHP_7_1;
         }
-        return self::LATEST_PHP_7_2;
+        if (version_compare($phpVersion, '7.3', 'lt')) {
+            return self::LATEST_PHP_7_2;
+        }
+        return self::LATEST_PHP_7_3;
     }
 
     /**